build: update unocss after template move
