+++ base/__init__.py
@@ -1,8 +1,8 @@
 """Expose core routers and constants used across the application."""
 
-from server.routes.ui.auth import router as auth_router
-from server.routes.ui.user_pages import router as user_pages_router
-from server.routes.ui.welcome import (
+from .routes.auth import router as auth_router
+from .routes.user_pages import router as user_pages_router
+from .routes.welcome import (
     router as dashboard_router,
     WELCOME_TEXT,
     INVENTORY_TEXT,
